% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ll}
\alias{ll}
\title{Example dataset, an epidemiological linelist from a treatment centre}
\format{
A data.frame with 12 rows and 10 variables:
\describe{
\item{id}{Patient identifier}
\item{site}{Site identifier}
\item{age}{Patient age in years}
\item{status}{Patient status}
\item{date_onset}{Date of symptom onset}
\item{date_admit}{Date of admission to treatment centre}
\item{date_lab}{Date of laboratory test}
\item{lab_result}{Result of laboratory test}
\item{date_exit}{Date of exit from treatment centre}
\item{outcome}{Patient outcome}
}
}
\usage{
ll
}
\description{
Example dataset, an epidemiological linelist from a treatment centre
}
\keyword{datasets}
