% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ll_queries}
\alias{ll_queries}
\title{Example set of queries to run on \code{ll} using \code{\link{query_vec}}}
\format{
A data.frame with 5 rows and 2 variables:
\describe{
\item{query_id}{Query IDs}
\item{query}{Query expressions in string format}
}
}
\usage{
ll_queries
}
\description{
Example set of queries to run on \code{ll} using \code{\link{query_vec}}
}
\keyword{datasets}
