% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_semicolon.R
\name{remove_semicolon}
\alias{remove_semicolon}
\title{Remove trailing semicolon from a string.}
\usage{
remove_semicolon(.x)
}
\arguments{
\item{.x}{a vector.}
}
\value{
A character vector with trailing semicolon removed.
}
\description{
Removes trailing semicolon from a string.
}
\examples{
remove_semicolon("This will be removed;")
}

