% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sll}
\alias{sll}
\title{Example dataset, a summary epidemiological linelist containing only
confirmed/probable cases}
\format{
A data.frame with 10 rows and 8 variables:
\describe{
\item{sll_id}{Patient identifier in the summary linelist}
\item{tc_admit}{Was patient admitted to a treatment centre?}
\item{tc_id}{Patient ID at treatment centre}
\item{tc_site}{Site of treatment centre}
\item{sll_age}{Date of symptom onset}
\item{sll_status}{Date of admission to hospital}
\item{sll_date_outcome}{Date of laboratory test}
\item{sll_outcome}{Patient outcome}
}
}
\usage{
sll
}
\description{
Example dataset, a summary epidemiological linelist containing only
confirmed/probable cases
}
\keyword{datasets}
