% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_whitespace.R
\name{remove_whitespace}
\alias{remove_whitespace}
\title{Remove excess white space from a character vector.}
\usage{
remove_whitespace(.x)
}
\arguments{
\item{.x}{a vector.}
}
\value{
A character vector with excess white space removed.
}
\description{
Remove excess white space from a character vector.
}
\examples{
remove_whitespace("  this is    too  much   white  space    ")
}

